Refactor email header and body helpers into dedicated modules
